feat:FRANKI LT changes in to game component
