+++ src/app/shared/const/plot.configuration.ts
@@ -1,0 +1,56 @@
+
+   export const plotConfigurationsConst = [
+      { planted: true, irrigated: true, cropIcon: '🌱' },
+      { planted: true, irrigated: false, cropIcon: '🌾' },
+      { planted: false, irrigated: false, cropIcon: '' },
+      { planted: true, irrigated: false, cropIcon: '🥕' },
+      { planted: false, irrigated: true, cropIcon: '' },
+      { planted: true, irrigated: true, cropIcon: '🌽' },
+      { planted: false, irrigated: false, cropIcon: '' },
+      { planted: true, irrigated: false, cropIcon: '🥬' },
+      
+      { planted: false, irrigated: false, cropIcon: '' },
+      { planted: true, irrigated: true, cropIcon: '🍅' },
+      { planted: true, irrigated: false, cropIcon: '🌱' },
+      { planted: false, irrigated: false, cropIcon: '' },
+      { planted: true, irrigated: false, cropIcon: '🥔' },
+      { planted: false, irrigated: true, cropIcon: '' },
+      { planted: false, irrigated: false, cropIcon: '' },
+      { planted: true, irrigated: false, cropIcon: '🌾' },
+      
+      { planted: true, irrigated: false, cropIcon: '🥕' },
+      { planted: false, irrigated: false, cropIcon: '' },
+      { planted: false, irrigated: true, cropIcon: '' },
+      { planted: true, irrigated: true, cropIcon: '🌽' },
+      { planted: false, irrigated: false, cropIcon: '' },
+      { planted: true, irrigated: false, cropIcon: '🥬' },
+      { planted: true, irrigated: false, cropIcon: '🍅' },
+      { planted: false, irrigated: false, cropIcon: '' },
+      
+      { planted: false, irrigated: false, cropIcon: '' },
+      { planted: true, irrigated: false, cropIcon: '🌱' },
+      { planted: true, irrigated: true, cropIcon: '🥔' },
+      { planted: false, irrigated: false, cropIcon: '' },
+      { planted: true, irrigated: false, cropIcon: '🌾' },
+      { planted: false, irrigated: true, cropIcon: '' },
+      { planted: false, irrigated: false, cropIcon: '' },
+      { planted: true, irrigated: false, cropIcon: '🥕' },
+      
+      { planted: true, irrigated: false, cropIcon: '🌽' },
+      { planted: false, irrigated: false, cropIcon: '' },
+      { planted: true, irrigated: true, cropIcon: '🥬' },
+      { planted: true, irrigated: false, cropIcon: '🍅' },
+      { planted: false, irrigated: false, cropIcon: '' },
+      { planted: false, irrigated: true, cropIcon: '' },
+      { planted: true, irrigated: false, cropIcon: '🌱' },
+      { planted: false, irrigated: false, cropIcon: '' },
+      
+      { planted: false, irrigated: false, cropIcon: '' },
+      { planted: true, irrigated: false, cropIcon: '🥔' },
+      { planted: true, irrigated: false, cropIcon: '🌾' },
+      { planted: false, irrigated: true, cropIcon: '' },
+      { planted: true, irrigated: true, cropIcon: '🥕' },
+      { planted: false, irrigated: false, cropIcon: '' },
+      { planted: true, irrigated: false, cropIcon: '🌽' },
+      { planted: true, irrigated: false, cropIcon: '🥬' }
+    ];
