+++ src/app/shared/const/slide.animation.ts
@@ -1,0 +1,11 @@
+import { trigger, style, transition, animate } from '@angular/animations';
+
+export const slideInAnimation = trigger('slideIn', [
+  transition(':enter', [
+    style({ transform: 'translateX(100%)', opacity: 0 }),
+    animate('0.5s ease-in-out', style({ transform: 'translateX(0)', opacity: 1 }))
+  ]),
+  transition(':leave', [
+    animate('0.5s ease-in-out', style({ transform: 'translateX(100%)', opacity: 0 }))
+  ])
+]);
